anticipation: refactor page file structure
